Feat/database connection (#3)
* feat: add more required packages

* feat: add connection string for postgresql datavase

* Add TaskDbContext for TaskModel management

Introduces a new `TaskDbContext` class inheriting from `DbContext` to manage `TaskModel` entities. Includes a `DbSet<TaskModel>` for CRUD operations and a constructor for configuration. The `OnModelCreating` method is overridden for future customization.

* Add Swagger and EF Core with PostgreSQL support

This commit introduces Swagger for API documentation by adding
`AddEndpointsApiExplorer()` and `AddSwaggerGen()` to the
service collection. It also configures Entity Framework to
use PostgreSQL by adding a DbContext for `TaskDbContext`
and setting up the connection string from the application's
configuration.